api/requirements.txt
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bandit==1.6.2
black==19.3b0
certifi==2020.12.5
chardet==4.0.0
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
decorator==4.4.0
Django==2.2.4
django-cors-headers==3.6.0
django-debug-toolbar==2.0
django-environ==0.4.5
django-extensions==2.2.1
django-filter==2.4.0
django-leaflet==0.27.1
django-workers==0.2.0
djangorestframework==3.10.2
djangorestframework-camel-case==1.2.0
djangorestframework-gis==0.16
drf-yasg==1.20.0
gitdb2==2.0.5
GitPython==3.0.2
gunicorn==19.9.0
idna==2.10
importlib-metadata==0.19
inflection==0.5.1
ipdb==0.12.2
ipython==7.7.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.15.1
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.1
parso==0.5.1
pbr==5.4.2
pexpect==4.7.0
pickleshare==0.7.5
pip-tools==4.1.0
pluggy==0.12.0
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.4.2
pylama==7.7.1
pyparsing==2.4.2
pytest==5.1.1
pytest-bandit==0.5.2
pytest-cov==2.7.1
pytest-django==3.5.1
pytz==2019.2
PyYAML==5.1.2
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.12.0
smmap2==2.0.5
snowballstemmer==1.9.0
sqlparse==0.3.0
stevedore==1.30.1
toml==0.10.0
traitlets==4.3.2
uritemplate==3.0.1
urllib3==1.26.2
wcwidth==0.1.7
Werkzeug==0.15.5
zipp==0.6.0
